move event pieces to separate file
